feat: build dropbear
We previously depended upon a custom version that was bundled with Trimui_Ex. This wasn't always guaranteed to exist, so compiling it in-house is the way to go.

Note that we use an older version of the golang image to avoid issues where libcrypt depends on XCRYPT_2.0, thus breaking the compiled dropbear package.